remove move argument in write for now
Pillow package has ImageFont that will return the length of a string in pixels. But it does not seem to work in Colab because Colab cannot access the font folder to determine the necessary font metrics. There might be a way to measure the pixel length of a string using the svg code, but need to research that.